kernel: disallow duplicate module names;
We did not handle the case of a module registering with the same name
as an already existing module. This caused the module count to be
wrong and the kernel to hang. Instead we panic now upon start if one
tries to register a module with the name of an existing module,
making the problem obvious.