coffin: refactored coffin type to contain a pointer to a tomb instead of a tomb;
A tomb contains a mutex which again may not be copied after it is
first used. If we copy a tomb we copy the mutex. However, if we got the
tomb from tomb.WithContext, another go routine already has a reference
to this tomb and thus will lock a different copy of the mutex than we
have stored in our coffin. Thus, if the tomb later tries to close the
dying channel the critical section of doing this might be entered twice
at the same time, causing a panic because we are closing a closed
channel.