ddb: fix parsing of json tags for ddb key and hash fields;
Suppose you have a struct like this:

type MyStruct struct {
    Field string `json:"field,omitempty" ddb:"global=hash"`
}

Previously gosoline would think the json tag for Field would be
"field,omitempty". This is however not what the programmer intended and
thus needs to resolve to "field" instead. Therefore this commit changes
the way we parse these tags to account for these kinds of tags.