clock: added ticker implementation;
This adds an improved ticker compared to the default ticker of go.
Instead of just ticking, it can also be reset, meaning the next tick
occurs only after the whole duration again. We are already using this
pattern in quite some places, but mostly with replacing the ticker
itself. Instead, you now can just call Ticker.Reset (which is even
thread-safe).