redis: flag redis.Nil as ErrorOk because it needs to be handled by the caller;
Redis returns `redis.Nil` as an error should no element be found. This
needs to be told to the executor to be okay because otherwise it retries
the action which produced the error. Instead, we need to signal this as
being a normal return value and let the caller of the action handle
this.