apiserver, cfg: fix parsing of config options and registration of custom validators;
The last update to the mapx/cfg package actually broke the parsing of
the api.port field as it saw that an int was convertible to a string by
using it as a code point. Thus, port 80 would get converted to "P"
instead of "80". This commit fixes this by only converting if we don't
change the kind of the value to avoid that issue.

We also ran into the problem that with the update of gin we could no
longer register custom validators (as gosoline was trying to cast
validator/v10 to validator/v8. We fixed that by requiring v10
validators, so this is a breaking change.