currency: added closeness margin;
When calling ToEurAtDate, ToUsdAtDate, ToCurrencyAtDate, HasCurrencyAtDate we need to supply a date parameter. If this date is in the future, these four methods will return an error. While logically this makes sense, in practice it causes the following problem: µService 1 creates an object A and then, right after, µService 2 receives this object and wants to convert its money related values. If the two microservices are running on different machines, it could be that object A has its "createdAt" field a few milliseconds/nanoseconds after the moment of time µService 2 receives it. In this scenario, µService 2 will receive a "date in the future" error from the currency module.

As a solution, if a call for any of the four above mentioned methods, has a date in the future which is less than a "closeness margin" (1 minute) apart, that call will succeed and it will use the "current/today" exchange rate.