test: added option to api test cases to specify the number of redirects to perform;
If you have a test case which should trigger a redirect from a service,
you might be more interested in having the correct redirect performed
than the response from the last redirect (because that is a page you
don't control or the protocol is not HTTP(S) anymore, but something like
`intent://`). You can now specify how many redirects to follow before
the last response is returned, defaulting to 0 (which may make some tests
which previously worked now fail, so you have to add the expected number
of redirects to these tests).