log: Do not write to map fields from handler as they should be read only;
The sentry handler is writing the event id to the logger fields so
subsequent log messages carry that event id. However, it was doing that
in an unsafe way as it simply modified the map. This is however not
allowed as the map could be read by another thread at the same time, for
example if that thread is currently executing `WithFields`. Instead, we
now build a completely new map with the sentry event id included. This
way we can't be writing to a map which is also read at the same time and
should be fine.