cloud/aws: replace aws default rate limiter with custom rate limiter;
The AWSv2 SDK by default uses a rate limiter with capacity for 100
retries before it stops retrying. A retry needs to be paid for by 5
successful requests (so your backoff needs to be long enough and your
general traffic good enough). So if you are suddenly hitting onto a
ddb table with very limited capacity, almost all requests fail and you
exhaust the pool of retries. So requests will fail even with an infinite
retry policy. To fix this, we disable the rate limiter and only use our
backoff delayer to limit the number of requests we send to AWS.