metric: allow defining custom units and added *Maximum and *Minimum units;
If you need to measure the maximum of something over time, you can do
that in cloudwatch, no problem. However, gosoline "helpfully" sums up
all your values before sending them to cloudwatch, making it impossible
to measure anything other than averages or sums. Thus, we define a
function to register your own custom units with the units they map to as
well as a function to reduce a set of values to a single value, and a
new set of units for maximum and minimum counts, seconds, and
milliseconds.