log: don't use the error message we constructed as a format string;
If you take some arbitrary message and use it as a format string, you
destroy the parts which randomly contain percent signs (url encoded data
for example). Thus, we need to supply a format string (just %s) on our
own to avoid this problem and get the message as we pass it in.