clock: added `NewTimer` and `NewTicker` functions to allow mocking more functionality and replaced clockwork;
When you need to wait for some time OR until the context expires, you
can't use `clock.After` (that would leak the timer created in the
background as you can't call `Stop` on a channel). Thus, you need to use
a `Timer`, but those are currently not mocked. This commit allows you to
mock those timers and control them from your unit test as needed.