cloud/aws/kinesis: do not consider finished shards or shards with an unfinished parent shard for reading;
If you change the number of shards in kinesis, you either split or
combine shards. Thus, for a limited amount of time, you might have a
different number of shards than you expect (e.g., when changing from 4
to 8 shards, there are 12 shards for some time, the old 4 shards and the
new 8 shards). We will now ignore the 4 shards you already finished (if
you are not behind on the stream) or any of the 8 new shards which have
still unfinished parent shards in the set of the old 4 shards when
assigning shards to clients. This also ensures you don't read data from
a child shard before completing the parent shard (which would violate
processing data with the same partition key in-order).