cloud/aws: fix context.Canceled error for larger S3 downloads; exec: added WithStoppableDeadlineContext;
When we request a file from S3, we do the initial request with a
deadline and, if successful, cancel the deadline. However, this then
cancels the context we need to download the body of that response -
causing spurious canceled errors when downloading larger files (smaller
files sometimes manage to slip through the window until the cancel is
detected). To fix this, we introduce a new context type which is stopped
but not canceled, so we can free the associated resources without
causing this problem.