stream: instead of marking the producer daemon as running, mark it as stopped;
When using the producer daemon, you could have your module start and
produce messages (mainly in tests) before the producer daemon was
running, causing calls to Write return an error. This commit changes the
behavior to only return an error when the producer daemon was already
shut down (and also makes reading that flag thread safe).