apiserver: added download handler (#897)
This release adds a new type of API handler to the apiserver package: The `DownloadHandler` will act like a stream handler but don't require passing a JSON body. The goal is to provide a possibility for the server to answer a clients GET request (i.e. from a HTML anchor) with a streamed response (i.e. a file).