ddb: report the consumed capacity again;
Our DynamoDB package used to report the capacity consumed by each
operation in the past. This stopped working at some point (not sure
when, maybe when switching to v2 of the AWS SDK?). This commit restores
this functionality and changes the API slightly. Getting the different
capacity figures now requires a function call instead of reading a
field, but adjusting code should not require more than a pair of
parenthesis. The change also guards against AWS being smart and not
reporting for example any consumed read capacity for a query, instead
just reporting total capacity (because a query can only ever read) by
providing the new function calls which take care of accessing the
correct field.