conc: Added task runner and scheduler modules;
The task runner (enabled by default) allows you to submit new tasks to
be run in a controlled environment. It should only be a method of last
resort to run go routines while still (somehow) tracking their results.
IF ANY OF THEM FAILS, THE WHOLE TASK RUNNER WILL BE SHUT DOWN.

This also adds a scheduler module (which currently uses the task runner
in the background) to batch requests for a specified time together (or
until reaching a batch size) and execute them all at once.