cloud/aws/kinesis: allow resuming consuming empty kinesis streams;
When a Kinesis stream doesn't get any messages for a day (on default
message expiration of one day), the last sequence number we have stored
becomes invalid and we effectively start consuming the stream from the
start each time our consumer restarts. To prevent this, we now also
store the last shard iterator in case we didn't receive any messages in
the last batch. Then, we can resume consuming from the shard iterator
and only fall as much behind as we took to restart the consumer.