log/mocks: Refactored logger mock factory by adding options and changed return value to interface;
Deprecated `NewLoggerMockedAll` and `NewLoggerMockedUntilLevel` and
replaced them with options to `NewLoggerMock` with the options
`WithMockUntilLevel` and `WithMockAll`. Also added the `WithTestingT`
option to supply a `testing.T` instance to the logger, allowing for
automatic expectation assertions and printing recorded log messages if a
test fails for easier debugging.

This commit also breaks backwards compatibility by changing the return
value of these methods to the new `LoggerMock` interface, so if you
store the mock in a field of a test suite or similar, you need to adjust
the type.